move ThreadPoolScheduler to parent package
